add function: get product for home and config category

+++ src/main/java/com/e_cormerce/shoppe/entity/product/Product.java
@@ -43,7 +43,7 @@ public class Product {
   float discount_percentage;
 
   @Column(nullable = false)
-  double total_quantity;
+  int total_quantity;
 
   @Enumerated(EnumType.STRING)
   @Column(
